+++ mark1/lib/pages/new_transaction/new_transaction.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:mymy_m1/helpers/templates/main_view_template.dart';
+import 'package:mymy_m1/helpers/templates/widget_templates.dart';
 
 class NewTransaction extends StatefulWidget {
   const NewTransaction({super.key});
@@ -10,6 +12,6 @@ class NewTransaction extends StatefulWidget {
 class _NewTransactionState extends State<NewTransaction> {
   @override
   Widget build(BuildContext context) {
-    return const Text("New");
+    return mainView(context, body: const CustomText(text: "New Transaction"));
   }
 }